Avoid replacing user page tables with kernel ones.
The unmap code would look at a page directory entry. If it was zeroed, it
would load the kernel page directory entry, assuming that in kernel mode
this might fill in a valid value, and for user mode addresses this would
always reset it to zero.

On x86 there is actually a page table at directory index zero, used for
processor bringup and sleep (eventually). We've never observed an issue
because 1) the process or heap is usually down within the first PDE
somewhere (if user mode creates its own page table there then the bug
won't hit), and 2) We don't usually call unmap on regions that aren't
previously mapped. Anyway, it's much safer to not install kernel PDEs
on user mode addresses, and ean easy check to make.