Store network address translation in sockets.
This change publishes the address translation structure and stores
translations in connected sockets. This will allow higher level protocols
like TCP to help determine whether or not an address translation is still
valid. ARP does not require its translation to timeout, but NDP does.

As part of this change, the ARP request and retry logic is moved out of
netcore and into the ARP module. Other address translation protocols may
have a different way of operating.