Convert link address entry state to an enum.
Link address entries were storing whether or not an address was configured
and, in another member, whether or not the address was static. This state
can be combined into an enum. This change is in preparation to make use of
the link address entry's state in NDP by adding values for 'tentative' and
'duplicate'.