MM: Better management of PTE Global bit.
In the course of investigating why the global bit causes odd TLB issues on
VMWare, we realized that the swap VA region probably didn't need the
global bit set. This led to a few other cleaup areas, such as mapping the
page tables created in the boot loader as global, and removing an
unnecessary memory barrier.

This change should not make any difference functionally or from a real
performance perspective, but is hygienic.