Create kernel stack before forked process.
In preparation for the next incoming change of using the self map for the
page directory, create the kernel stack for the forked thread before the
process. Reasoning below.

Before, when each page directory had its own separate mapping in kernel
space, poking another process' page directory was easy. Now, because we're
about to use the self map for the page directory, mappings to the page
directory of another process are more expensive. We have to raise to
dispatch, map it in the swap page, then unmap and invalidate that TLB
entry (a serializing operation).

PspCreateThread inside fork was one of the regular consumers of this
poking action. It would create a new thread for the child process, but
then poked its page directory because it needed to ensure the kernel stack
of its new thread was viable on the new process before it could be run.

In order to avoid the expensive map-and-poke operation, especially when
99% of new threads don't even need it, this change is reordering
operations a bit. If we allocate the kernel stack *before* creating the
page directory for the new process, then we know that when the new
process' page directory is initialized with the kernel PDEs, it will have
the most up to date version, including our new stack. So simply by moving
the order around, we can avoid that (now expensive) poke of the PDE
altogether.

This made things a little more awkward, as PspCreateProcess needs to
accept a pre-allocated kernel stack. In the other (non-fork) calls to
PspCreateThread, where no kernel stack is provided, the poke still needs
to occur. For threads destined for the same process they were created
from, this is cheap, since the self map can be used. The only other place
where this is not the case is PsCreateProcess, creating the
PspLoaderThread for the very first process. This still needs the expensive
version of the poke, but is very cold code.