x86: Use self map to workaround VMWare TLB bug
This change restructures the x86 mapping code to use the self map rather
than an explicit mapping when referencing the page directory. This is
done to work around what seems like a bug in VMWare on Windows,
explained below.

Symptom: Page faults inside the self map region (ie when touching a page
table), even though we explicitly check that a PDE is present before
attempting to touch that page table. These faults do not occur on real
hardware nor on VirtualBox, but occur reliably on VMWare Workstation
12.5.1-12.5.7, currently the latest version.

Problem: TLB mismanagement somewhere in the hypervisor. I know this
because immediately after the fault I can copy the contents of the page
directory accessed via the self map and see the kernel's page directory,
even though CR3 is set to some other process' page directory. If I
invalidate the TLB by reloading CR3, then suddenly via the self map the
page directory is correct. The PDE for the self map is not marked global
in either process, nor is any PDE for user mode page tables, which are
always the page tables we're trying to touch when we fault.

On real hardware this simply could not happen. With the PDEs set up how
they are, reloading CR3 would cause a flush of the TLB for the entire
region of the self map covering user mode. Additionally, the hardware
doesn't cache negative TLB entries. So you might expect something that
should fault to not fault and go somewhere weird, but you should never get
a case where a valid mapping faults. It's like parts of the TLB or paging
structure cache are not being cleared on CR3 switch. This VM has 4 cores,
on a relatively quiet 8-core physical host. I wondered for awhile whether
it had to do with rescheduling the vCPU on other cores, and not properly
clearing out all state from the old core. I could never quite build out a
mental scenario for that.

Here's where things get interesting: if I turn off all global mappings and
never set the G bit in a PDE/PTE, the problem disappears. I spent the next
day narrowing down exactly which G bit mapping causes the problem. It
turns out to be our extra mapping of the page directory. We use a block
allocator, which means page directories are reused, and get their own
private mappings like any heap region would. This was convenient for us
because operations like map-in-other-process were one step easier.

My theory is that the global mapping to the page directory was somehow
percolating back where it shouldn't have. If translation is a two-step
process of 1) Guest VA-to-PA, and 2) Guest PA to host PA, and the Global
bit was somehow saved in that second step, then maybe it was percolating
inappropriately back up into the self-map PDE somehow, even though that
PDE is not Global. That might explain why the self map PD is bad, but
doesn't explain why user mode PTs referenced through the self map faulted.
Now that I know the problem, I gave up speculating on how their hypervisor
works.

Our fix is to use the self map to access the page directory, which is a
cleanup item I'd been meaning to do anyway, and stop mapping the page
directory elsewhere with global attributes. A couple of additional changes
fell out of the fact that accessing another processes page directory is
now more expensive.

The big one was this business during context switching of ensuring that
the old stack is present in the new process. This is harder now that the
new process PD is tougher to poke. We were doing this because our order
was 1) Switch CR3, 2) set CurrentThread, and then 3) Switch to new stack.

A much easier order would be 1) Set CurrentThread, 2) Switch to new stack,
then 3) Switch CR3. This was problematic before because if any structure
faulted due to a missing PDE, the CheckForDirectoryUpdates code would look
in CurrentThread to get the page directory, and end up with the wrong one.
Now that the self map is in use, directory updates can be fixed with
nothing but the stack.

In order to make that work on ARM, where I need the explicit mappings to
fake up the self map, I had to save the VA of the first level table in the
processor block somewhere. This is simply to avoid reference through
CurrentThread->OwningProcess->AddressSpace, each of which might cause an
infinite fault loop. I saved it in the TSS slot, which was just wasting
space on ARM anyway. Eventually we should refactor the
architecture-specific stuff in the processor block, and maybe put it just
before the main structure.

The last tricky bit is ensuring that the new stack is accessible from the
old CR3, since we now switch stacks and then change CR3. This turns out to
be pretty easy if you just dereference the stack before switching to it in
the context swap assembly.

So, a lot of changes fell out of a workaround, but they probably put us
more in line with the common and well-tested path, which is good for
avoiding issues like this.