Debugui: Set a sane default font size.
The yHeight member of CHARFORMAT2 is specified in 1/20th of a point size
increments. We were requesting a very tiny default font size. Most
machines were ignoring that and picking a sane default. A few machines can
actually render that size, leading to a very tiny debugger display.

Also fixed a typo found in a comment while brushing up on TCP.