lib: fatlib: Avoid unaligned access to LFN name fields
GCC 9 now starts emitting warnings (turned errors) because we grab the
address of the Long File Name FAT directory entries, which are PUSHORTs,
but are not aligned. We then access those pointers as an array. We do
(at least in all cases that mattered) use the READ_UNALIGNED16() macro,
but GCC isn't smart enough to see that we're careful with how we use the
pointer.

To fix, convert to a char pointer, and access with care, as before.

Also, FAT_DIRECTORY_ENTRY structures are always aligned to 32-byte
boundaries, so mark their structs as such if it helps with other fields.